requirements.txt
base58==2.1.1
bip-utils==2.10.0
tonsdk==1.0.15
cryptography==46.0.3
pyfiglet==1.0.4
PyNaCl==1.5.0
questionary==2.1.1
requests==2.32.5
rich==14.2.0
substrate-interface==1.7.11
polyseed @ git+https://github.com/DiosDelRayo/polyseed-python.git
reportlab>=4.0
qrcode[pil]>=7.4.2
shamirs>=4.0.0